Rename to udata-schema-gouvfr (#1)
